second

+++ 037_array_subseq/maxSeq.c
@@ -5,7 +5,7 @@ size_t maxSeq(int * array, size_t n) {
   }
   size_t sum = 1;
   size_t subseq = 1;
-  for (int i = 0; i < n - 1; i++) {
+  for (size_t i = 0; i < n - 1; i++) {
     if (array[i] < array[i + 1]) {
       subseq++;
     }
